Module car_demo::build

Builds the car as a series of joints.

Functions

add_car_mesh ğŸ”’

Adds car mesh.

add_engine ğŸ”’

Basic add engine function that adds an engine to an entity.

add_tire_contact ğŸ”’

Adds a tire contact to an entity.

add_wheel_scene ğŸ”’

Adds wheel meshes.

build_chassis ğŸ”’

Builds the chassis from a series of joints.

build_model

Builds the car model, utilizing other functions to create the chassis, suspensions, and wheels.

build_steer ğŸ”’

Builds steering system. Similar to build_suspension, but with an rz joint, and no mesh and no contact.

build_suspension ğŸ”’

Builds suspension.

build_wheel ğŸ”’

Builds the wheels.