<<<<<<< HEAD

Crate car_demo

This is our driving simulation for our 2023 Capstone project. It’s an example of how engineering simulations can be created with Rust ======= <<<<<<< HEAD

Crate car_demo

Modules

build 🔒

camera_az_el 🔒

controller 🔒

enviornment 🔒

physics

schedule 🔒

Structs

FpsText 🔒

RPMText

SpeedText

Constants

FIXED_TIMESTEP 🔒

Functions

create_ui 🔒

fps_updater 🔒

main 🔒

setup_system

=======

Crate car_demo

This is our driving simulation for our 2023 Capstone project. It’s an example of how engineering simulations can be created with Rust >>>>>>> c2d6dc236c53139cb1631495f4ee8dff8b3bb18c and the game engine Bevy. See README for more information about compiling, running, and using this project. Use this documentation to learn more about individual functions and the structure of the program.

Modules

build 🔒

Builds the car as a series of joints.

camera_az_el 🔒

enviornment 🔒

Builds a simple environment with lighting.

physics

Contains all driving simulation specific physics systems as well as system to handle controller and keyboard input. Also contains systems to simulate an engine with a gear box.

schedule 🔒

Contains the physics schedule, which determines which systems run in which order.

Structs

FpsText 🔒

Fps Text Struct

RPMText

RPM Text Struct

SpeedText

Speed Text Struct

Constants

FIXED_TIMESTEP 🔒

Set a larger timestep if the animation lags

Functions

create_ui 🔒

Creates and updates the UI.

fps_updater 🔒

Updates the text which displays the current FPS.

main 🔒

Main function. Creates app, sets up window, adds startup systems that build the car and environment, and specifies the order of other systems including the <<<<<<< HEAD physics schedule and UI.

setup_system

Sets up the UI, the environment, and the model.

======= physics schedule and UI.

setup_system

Sets up the UI, the environment, and the model.

>>>>>>> 4fc876cc0cc631b86384b38256300d0e9ad03e43 >>>>>>> c2d6dc236c53139cb1631495f4ee8dff8b3bb18c