<<<<<<< HEAD

Module car_demo::build

Builds the car as a series of joints.

Functions

add_car_mesh 🔒

Adds car mesh.

add_engine 🔒

Basic add engine function that adds an engine to an entity.

add_tire_contact 🔒

Adds a tire contact to an entity.

add_wheel_scene 🔒

Adds wheel meshes.

build_chassis 🔒

Builds the chassis from a series of joints.

build_model

Builds the car model, utilizing other functions to create the chassis, suspensions, and wheels.

build_steer 🔒

Builds steering system. Similar to build_suspension, but with an rz joint, and no mesh and no contact.

build_suspension 🔒

Builds suspension.

build_wheel 🔒

Builds the wheels.

======= <<<<<<< HEAD

Module car_demo::build

Functions

add_car_mesh 🔒

spawns car chassis into scene and apply transformations

add_engine 🔒

basic add engine function that could be expanded and customized

add_tire_contact 🔒

creates stiffness and damping of suspension systems

add_wheel_scene 🔒

used in build_wheel() to spawn wheels to scene and apply transformations to each wheel

build_chassis 🔒

build the chassis from a series of joints

build_model

initializes 3rd person camera, pulls 3D models from assets folder, sets coordinates for model transformations, creates parent hierarchy (chassis –> suspension –> wheels)

build_steer 🔒

similar to build_suspension, but with an rz joint, and no mesh and no contact

build_suspension 🔒

creates RBDA joints to connect wheel to cars (allows the car to lean to the side fron g-force simulation)

build_wheel 🔒

assigns wheel models into the scene to their parented suspension

=======

Module car_demo::build

Builds the car as a series of joints.

Functions

add_car_mesh 🔒

Adds car mesh.

add_engine 🔒

Basic add engine function that adds an engine to an entity.

add_tire_contact 🔒

Adds a tire contact to an entity.

add_wheel_scene 🔒

Adds wheel meshes.

build_chassis 🔒

Builds the chassis from a series of joints.

build_model

Builds the car model, utilizing other functions to create the chassis, suspensions, and wheels.

build_steer 🔒

Builds steering system. Similar to build_suspension, but with an rz joint, and no mesh and no contact.

build_suspension 🔒

Builds suspension.

build_wheel 🔒

Builds the wheels.

>>>>>>> 4fc876cc0cc631b86384b38256300d0e9ad03e43 >>>>>>> c2d6dc236c53139cb1631495f4ee8dff8b3bb18c